Add sites support (should be improvement over FreeBSD ports)
Proof of concept of new approach to expanding DL_SITE (aka MASTER_SITE)
keywords.  The FreeBSD approach is to use make logic extensively, on
every call, when it's not used 99% of the time.  By converting this to
a shell script, we get the following benefits:
 1) cleaner (IMO)
 2) should be much faster
 3) only evaluated during fetch attempt
 4) use of shell case/esac ensures only one keyword is expanded
    (basically double-booking reason #2)

I only added a few big keywords like SOURCEFORGE/SF and APACHE
family.  The rest will be added as new Ravenports require.