Add install-shifting and gmake port (2 variants)
To force install during the build phase (where the toolchain is still
hooked in), use INSTALL_REQ_TOOLCHAIN=yes in the makefile.

The gmake port has two variants, standard and lite.  The lite one
will be used to build other ports as the standard one has a dependency
chain to untangle.